requirements.txt
amazon-braket-default-simulator==1.26.2
amazon-braket-schemas==1.22.4
amazon-braket-sdk==1.88.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.2
backoff==2.2.1
backports.entry-points-selectable==1.3.0
boltons==24.1.0
boto3==1.35.81
botocore==1.35.81
cffi==1.17.1
cloudpickle==2.2.1
cryptography==44.0.0
importlib_metadata==8.5.0
jmespath==1.0.1
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==2.0.2
openpulse==1.0.1
openqasm3==1.0.0
opt_einsum==3.4.0
oqpy==0.3.7
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
s3transfer==0.10.4
scipy==1.13.1
six==1.17.0
sympy==1.13.3
typing_extensions==4.12.2
urllib3==1.26.20
zipp==3.21.0
